+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# 0.23.0
 - [BREAKING] Don't display `aria-pressed` when the component is opened. The attribute is not
   present by default, but can be bound from the outside.
 - [BREAKING] Don't display `aria-haspopup` by default, but display it if passed in a truthy value.